distributed-system/traffic-analytics/traffic-analytics-service/requirements.txt
grpcio==1.59.0
prometheus_client==0.15.0
protobuf==4.25.0
psycopg2==2.9.9
python_dateutil==2.8.2
Requests==2.31.0
